requirements.txt
flask
ultralytics
opencv-python
numpy
python-chess 